Set up MainContainer and grid system in container components
