+++ src/frontend/Containers/MainContainer.tsx
@@ -1,0 +1,15 @@
+import Container from '@material-ui/core/Container';
+import Grid from '@material-ui/core/Grid';
+import { SideBarContainer } from './SideBarContainer';
+import { StateTreeContainer } from './StateTreeContainer';
+
+export const MainContainer = (): JSX.Element => {
+    return (
+        <Container>
+            <Grid container>
+                <SideBarContainer />
+                <StateTreeContainer />
+            </Grid>
+        </Container>
+    );
+}
