+++ src/frontend/Containers/StateTreeContainer.tsx
@@ -1,7 +1,9 @@
+import Grid from '@material-ui/core/Grid';
+
 export const StateTreeContainer = (): JSX.Element => {
     return (
-        <div>
+        <Grid container item>
             State Tree Goes here
-        </div>
+        </Grid>
     );
 }
